James Clessuras Drafter/requirements.txt
Flask>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
Werkzeug>=2.3.0
supabase>=2.0.0
python-dotenv>=1.0.0 